Improve auth redirect abstraction
- Added initiate_login helper to build default session
- Move get_login_url into sentry.utils.auth
- Ensure 2fa is reset upon login initialization
