+++ src/sentry/web/decorators.py
@@ -6,7 +6,8 @@ from django.http import HttpResponseRedirect
 from django.contrib import messages
 from django.utils.translation import ugettext_lazy as _
 
-from sentry.web.helpers import render_to_response, get_login_url
+from sentry.utils import auth
+from sentry.web.helpers import render_to_response
 
 ERR_BAD_SIGNATURE = _('The link you followed is invalid or expired.')
 
@@ -15,12 +16,12 @@ def login_required(func):
     @wraps(func)
     def wrapped(request, *args, **kwargs):
         if not request.user.is_authenticated():
-            request.session['_next'] = request.get_full_path()
+            auth.initiate_login(request, next_url=request.get_full_path())
             if 'organization_slug' in kwargs:
                 redirect_uri = reverse('sentry-auth-organization',
                                        args=[kwargs['organization_slug']])
             else:
-                redirect_uri = get_login_url()
+                redirect_uri = auth.get_login_url()
             return HttpResponseRedirect(redirect_uri)
         return func(request, *args, **kwargs)
     return wrapped
@@ -32,7 +33,7 @@ def signed_auth_required(func):
         if not request.user_from_signed_request:
             messages.add_message(
                 request, messages.ERROR, ERR_BAD_SIGNATURE)
-            return HttpResponseRedirect(get_login_url())
+            return HttpResponseRedirect(auth.get_login_url())
         return func(request, *args, **kwargs)
     return wrapped
