+++ src/sentry/web/frontend/organization_member_settings.py
@@ -8,9 +8,9 @@ from django.utils.translation import ugettext_lazy as _, ugettext
 from sentry import roles
 from sentry.models import OrganizationMember, OrganizationMemberTeam, \
     Team, TeamStatus
+from sentry.utils import auth
 from sentry.web.frontend.base import OrganizationView
 from sentry.web.forms.edit_organization_member import EditOrganizationMemberForm
-from sentry.web.helpers import get_login_url
 
 
 class OrganizationMemberSettingsView(OrganizationView):
@@ -68,7 +68,7 @@ class OrganizationMemberSettingsView(OrganizationView):
                 id=member_id,
             )
         except OrganizationMember.DoesNotExist:
-            return self.redirect(get_login_url())
+            return self.redirect(auth.get_login_url())
 
         if request.POST.get('op') == 'reinvite' and member.is_pending:
             return self.resend_invite(request, organization, member)
